+++ src/neural_network/layer/layer_weight.rs
@@ -16,6 +16,7 @@ use ndarray::prelude::*;
 /// - `Conv3D` - Contains weights for 3D convolutional layers
 /// - `BatchNormalization` - Contains weights for batch normalization layers
 /// - `LayerNormalizationLayer` - Contains weights for layer normalization layers
+/// - `InstanceNormalizationLayer` - Contains weights for instance normalization layers
 /// - `Empty` - Represents a layer with no trainable parameters
 pub enum LayerWeight<'a> {
     Dense(DenseLayerWeight<'a>),
@@ -29,6 +30,7 @@ pub enum LayerWeight<'a> {
     Conv3D(Conv3DLayerWeight<'a>),
     BatchNormalization(BatchNormalizationLayerWeight<'a>),
     LayerNormalizationLayer(LayerNormalizationLayerWeight<'a>),
+    InstanceNormalizationLayer(InstanceNormalizationLayerWeight<'a>),
     Empty,
 }
 
@@ -195,6 +197,17 @@ pub struct BatchNormalizationLayerWeight<'a> {
 /// - `gamma` - Scale parameter (learned during training) that controls the variance of normalized values
 /// - `beta` - Shift parameter (learned during training) that controls the mean of normalized values
 pub struct LayerNormalizationLayerWeight<'a> {
+    pub gamma: &'a ArrayD<f32>,
+    pub beta: &'a ArrayD<f32>,
+}
+
+/// Weights for an instance normalization layer
+///
+/// # Fields
+///
+/// - `gamma` - Scale parameter (learned during training) that controls the variance of normalized values
+/// - `beta` - Shift parameter (learned during training) that controls the mean of normalized values
+pub struct InstanceNormalizationLayerWeight<'a> {
     pub gamma: &'a ArrayD<f32>,
     pub beta: &'a ArrayD<f32>,
 }