+++ src/test/neural_network_test.rs
@@ -23,6 +23,7 @@ mod global_max_pooling_1d_test;
 mod global_max_pooling_2d_test;
 mod global_max_pooling_3d_test;
 mod gru_test;
+mod instance_normalization_test;
 mod layer_normalization_test;
 mod layer_weight_test;
 mod loss_function_test;